+++ Addon/ElvUI_mMediaTag/core/changelog/3.60.lua
@@ -1,0 +1,26 @@
+mMT.Changelog[360] = {
+	DATE = "04.10.2024",
+	NEW = {
+		"[Class-icons]: Add Class icons to details and optimized the functions",
+		"[Datatext-Durability]: Add Summon repair mount function.",
+		"[Details-Embedded]: Details embedded feature for up to 4 Details Windows.",
+		"[Portraits]: Add desaturate feature if Unit is dead, default is off.",
+	},
+	UPDATE = {
+		"[Dock-Character]: Update tooltip.",
+		"[Health-Markers]: Update for TWW S1, thx to Trenchy.",
+		"[Portraits]: Update Zoom function, only one Zoom for All Styles.",
+		"[System]: DropDown Menu uses now the ElvUI Font & Fontflag.",
+		"[System]: Update TOC for Classic.",
+		"[Teleports]: Add Wormhole Generator: Khaz Algar's, thx to Cliaux.",
+		"[Teleports]: Update Teleports for TWW S1.",
+	},
+	FIX = {
+		"[Datatext-Score]: Prevent nil Errors with the new Season of TWW.",
+		"[Important-Spells]: Prevent nil Errors with not existing DB.",
+		"[ObjectiveTracker]: Fixed Show only Quests on Map function, thx to Cliaux.",
+		"[Portraits]: Changed Boss check priority, it should now work as intended.",
+		"[Portraits]: Possible fix for not showing Portrait if Player is in a vehicle.",
+		"[Portraits]: Wrong Portraits when Profile switching.",
+	},
+}