+++ Addon/ElvUI_mMediaTag/init.lua
@@ -35,7 +35,7 @@ mMT.firstLoad = 0
 mMT.Classes = { "DEATHKNIGHT", "DEMONHUNTER", "DRUID", "EVOKER", "HUNTER", "MAGE", "MONK", "PALADIN", "PRIEST", "ROGUE", "SHAMAN", "WARLOCK", "WARRIOR" }
 mMT.Locales = LibStub("AceLocale-3.0"):GetLocale("mMediaTag")
 mMT.Changelog = {}
-mMT.classIcons = {}
+mMT.ClassIcons = {}
 mMT.BossIDs = {}
 
 mMT.Modules.Portraits = {}
@@ -317,6 +317,8 @@ function mMT:PLAYER_ENTERING_WORLD(event)
 		end
 	end
 
+	if IsAddOnLoaded("Details") then mMT:SetupDetails() end
+
 	-- Modules
 	UpdateModuleSettings()
 	UpdateModules()
@@ -358,7 +360,7 @@ function mMT:CHALLENGE_MODE_START()
 end
 
 function mMT:ZONE_CHANGED_NEW_AREA()
-	mMT.Modules.ObjectiveTracker:TrackUntrackQuests()
+	if E.db.mMT.objectivetracker.settings.zoneQuests then mMT.Modules.ObjectiveTracker:TrackUntrackQuests() end
 	mMT:UpdateText()
 end