+++ Addon/ElvUI_mMediaTag/modules/nameplates/healthandexecute.lua
@@ -4,6 +4,14 @@ local NP = E:GetModule("NamePlates")
 local LSM = LibStub("LibSharedMedia-3.0")
 
 local HM_NPCs = {
+	-------- TWW Dungeons --------
+	[129208] = { 66, 33 },
+	[207940] = { 50 },
+	[210797] = { 44.5 },
+	[211087] = { 50 },
+	[213937] = { 60 },
+	[219984] = { 50 },
+
 	-------- DF Dungeons --------
 	-- Azure Vault
 	[186738] = { 75, 50, 25 }, -- Umbrelskul